backend/requirements.txt
flask
flask-cors
youtube-transcript-api
langchain
langchain-google-genai
langchain-community
faiss-cpu
python-dotenv
requests